MAINT/REF: isolate grafana log display and allow us to disable it
